small update to scripts

+++ compile_sources.sh
@@ -1,8 +1,17 @@
 #!/bin/bash
 
+while getopts "r" opt; do
+  case $opt in
+    r) recompile=true
+    ;;
+    \?) echo "Invalid option -$OPTARG" >&1; exit 1;
+    ;;
+  esac
+done
+
 # (re)compile Sylvan
 cd sylvan
-rm -r -f build
+if [[ $recompile ]]; then rm -r -f build; fi
 mkdir -p build
 cd build
 cmake ..
@@ -12,7 +21,7 @@ cd ..
 
 # (re)compile reachability code
 cd reach_algs
-rm -r -f build
+if [[ $recompile ]]; then rm -r -f build; fi
 mkdir -p build
 cd build
 cmake ..
